Refactor card presenter into smaller modules
